change int weights to double weights

+++ DotNet/Domain/Models/TileWeight.cs
@@ -2,13 +2,13 @@ namespace Domain.Models
 {
     public struct TileWeight
     {
-        public TileWeight(TileType tileType, int weight)
+        public TileWeight(TileType tileType, double weight)
         {
             TileType = tileType;
             Weight = weight;
         }
         
         public TileType TileType { get; }
-        public int Weight { get; }
+        public double Weight { get; }
     }
 }
